refactor: test/calcultor/model로 이동
